Add files via upload
The .py file that I'm working on, with reference to Ch 2-3 of the NLTK Book, and with helpful tips that I've extrapolated from the 

MIT course : '6.863J/9.611J Natural Language Processing: Fall 2012' under Prof. Robert Berwick. 

Link to course : http://web.mit.edu/6.863/www/fall2012/

This is for me to learn, but if it helps anyone jumpstart their NLP basics, by all means.